# backend/docs/api-spec.yaml
openapi: 3.0.0
info:
  title: Felparkering API
  description: Rest API för att hantera felparkerade fordon
  version: '1.0'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    variables: {}
paths:
  /login:
    post:
      tags:
        - Auth
      summary: Login user
      description: |-
        ### POST /login

        Loggar in användare med e-post och lösenord.  
        Returnerar en JWT-token [`sub`: User ID, `claim`: User Role] som används för autentisering.
      operationId: Loginuser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: abc123
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                data:
                  token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
                message: User logged in successfully
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation error
                  value:
                    errors:
                      - field: email
                        message: Invalid email address
                missingCredentials:
                  summary: Missing credentials
                  value:
                    error:
                      code: BAD_REQUEST
                      message: Missing credentials
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid credentials
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: User not found
      deprecated: false
      security: []
  /register:
    post:
      tags:
        - Auth
      summary: Register new user
      description: |-
        ### POST /register

        Registrerar en användare (CUSTOMER) med e-post och lösenord.
        Returnerar en JWT-token [`sub`: User ID, `claim`: User Role] som används för autentisering.
      operationId: Registernewuser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: admin@example.com
                password:
                  type: string
                  example: abc123
                confPassword:
                  type: string
                  example: abc123
      responses:
        '201':
          description: Created
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                  data: 
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
                  message: User registered successfully
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ValidationErrorResponse'
                  - $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation error
                  value:
                    errors:
                      - field: email
                        message: Invalid email address
                missingCredentials: 
                  summary: Missing credentials
                  value:
                    error:
                      code: BAD_REQUEST
                      message: Missing credentials
                passwordMismatch:
                  summary: Password and confirmation does not match
                  value:
                    error:
                      code: BAD_REQUEST
                      message: Password and confirmation does not match
        '409':
          description: Conflict
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: CONFLICT
                  message: User already exists
      deprecated: false
...
  /addresses/search:
    get:
      tags:
        - Address
      summary: Search addresses
      description: |-
        ### GET /addresses/search

        Söker adresser baserat på gatunamn och returnerar ett förslag per husnummer.
      operationId: Searchaddresses
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressSearchResponse'
              example:
                data:
                  - id: 101
                    street: Södra vägen
                    city: Stockholm
                    houseNumber: "1"
                  - id: 101
                    street: Södra vägen
                    city: Stockholm
                    houseNumber: "3"
                message: Addresses fetched successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
      deprecated: false
      security:
        - bearer: []
  /addresses/route:
    post:
      tags:
        - Address
      summary: Get route between two coordinates
      description: |-
        ### POST /addresses/route

        Hämtar en körbar rutt mellan två koordinatpunkter via extern route service.

        `start` och `end` anges som `[latitude, longitude]`.
      operationId: Getroute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RouteRequest'
            example:
              start:
                - 59.3293
                - 18.0686
              end:
                - 59.3326
                - 18.0649
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteApiResponse'
              example:
                data:
                  type: FeatureCollection
                  features:
                    - type: Feature
                      properties:
                        summary:
                          distance: 1200.5
                          duration: 320.2
                      geometry:
                        type: LineString
                        coordinates:
                          - [18.0686, 59.3293]
                          - [18.0649, 59.3326]
                message: Route fetched successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: Access Denied
      deprecated: false
      security:
        - bearer: []
  /admin/users:
    post:
      tags:
        - Admin
      summary: Create user
      description: |-
        ### POST /admin/users

        Skapar en användare, till exempel en parkeringsvakt.

        #### Behörighet:

        - `ADMIN`: Ja
            
        - `ATTENDANT`: Nej (403)
            
        - `CUSTOMER`: Nej (403)
      operationId: Createuser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUserCreateRequest'
            example:
              email: vakt@example.com
              password: secret123
              role: ATTENDANT
              attendantGroup:
                id: 1
                name: Stockholm Södra
      responses:
        '201':
          description: Created
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserCreatedResponse'
              example:
                data:
                  id: 7
                  role: ATTENDANT
                message: User was created successfully
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: BAD_REQUEST
                  message: Invalid request parameter or value
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: Access Denied
      deprecated: false
      security:
        - bearer: []
    get:
      tags:
        - Admin
      summary: Get all users
      description: |-
        ### GET /admin/users

        Hämtar en lista med alla användare.

        #### Behörighet:

        Roll:

        - `ADMIN`: ser alla användare
            
        - `ATTENDANT`: ej tillåten (403)
            
        - `CUSTOMER`: ej tillåten (403)
      operationId: Getallusers
      parameters: []
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUsersResponse'
              example:
                data:
                  - id: 1
                    role: CUSTOMER
                  - id: 7
                    role: ATTENDANT
                message: Users fetched successfully
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: Access Denied
      deprecated: false
      security:
        - bearer: []
  /admin/users/{id}:
    delete:
      tags:
        - Admin
      summary: Delete user
      description: |-
        ### DELETE /admin/users/{id}

        Tar bort en användare med {id}.

        #### Behörighet:

        - `ADMIN`: Ja
            
        - `ATTENDANT`: ej tillåtet (403)
            
        - `CUSTOMER`: ej tillåtet (403)
            

        #### Parameter

        - `id` _(Long)_ – användarens ID
      operationId: Deleteuser
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletedUserResponse'
              example:
                data:
                  id: 7
                message: User deleted successfully
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: Access Denied
        '404':
          description: Not Found
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: NOT_FOUND
                  message: User not found
      deprecated: false
      security:
        - bearer: []
  /admin/attendants:
    get:
      tags:
        - Admin
      summary: Get all Attendant Groups
      description: |-
        ### GET /admin/attendants

        Hämtar en lista med alla vaktgrupper.

        #### Behörighet:

        Roll:

        - `ADMIN`: ser alla vaktgrupper
            
        - `ATTENDANT`: ej tillåten (403)
            
        - `CUSTOMER`: ej tillåten (403)
      operationId: GetallAttendantGroups
      parameters: []
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminAttendantGroupsResponse'
              example:
                data:
                  - name: Stockholm Södra
                    attendants:
                      - id: 7
                        role: ATTENDANT
                  - name: Stockholm Norra
                    attendants: []
                message: Attendant groups fetched successfully
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: UNAUTHORIZED
                  message: Missing or invalid token
        '403':
          description: Forbidden
          headers: {}
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error:
                  code: ACCESS_DENIED
                  message: Access Denied
      deprecated: false
      security:
        - bearer: []
components:
  schemas:
    LoginResponse:
      title: LoginResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserDetailDto'
        message:
          type: string
      example:
        data:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        message: User logged in successfully
    ValidationErrorResponse:
      title: ValidationErrorResponse
      required:
        - errors
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
              message:
                type: string
      example:
        errors:
          - field: status
            message: must not be null
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
      example:
        error:
          code: BAD_REQUEST
          message: Bad request
    ErrorBody:
      title: ErrorBody
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
      example:
        code: BAD_REQUEST
        message: Bad request
    RegisterResponse:
      title: RegisterResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserDetailDto'
        message:
          type: string
      example:
        data:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
        message: User registered successfully
...
    AttendantGroupDto:
      title: AttendantGroupDto
      required:
        - id
        - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
      example:
        id: 1
        name: Stockholm Södra
...
    AddressSuggestion:
      title: AddressSuggestion
      required:
        - id
        - street
        - city
        - houseNumber
      type: object
      properties:
        id:
          type: integer
          format: int64
        street:
          type: string
        city:
          type: string
        houseNumber:
          type: string
      example:
        id: 101
        street: Södra vägen
        city: Stockholm
        houseNumber: "1"
    AddressSearchResponse:
      title: AddressSearchResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AddressSuggestion'
        message:
          type: string
      example:
        data:
          - id: 101
            street: Södra vägen
            city: Stockholm
            houseNumber: "1"
        message: Addresses fetched successfully
    RouteRequest:
      title: RouteRequest
      required:
        - start
        - end
      type: object
      properties:
        start:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: double
          description: Coordinate pair as [latitude, longitude].
        end:
          type: array
          minItems: 2
          maxItems: 2
          items:
            type: number
            format: double
          description: Coordinate pair as [latitude, longitude].
      example:
        start:
          - 59.3293
          - 18.0686
        end:
          - 59.3326
          - 18.0649
    RouteResponse:
      title: RouteResponse
      type: object
      description: GeoJSON route response proxied from OpenRouteService.
      additionalProperties: true
      properties:
        type:
          type: string
        features:
          type: array
          items:
            type: object
            additionalProperties: true
      example:
        type: FeatureCollection
        features:
          - type: Feature
            properties:
              summary:
                distance: 1200.5
                duration: 320.2
            geometry:
              type: LineString
              coordinates:
                - [18.0686, 59.3293]
                - [18.0649, 59.3326]
    RouteApiResponse:
      title: RouteApiResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/RouteResponse'
        message:
          type: string
      example:
        data:
          type: FeatureCollection
          features:
            - type: Feature
              properties:
                summary:
                  distance: 1200.5
                  duration: 320.2
              geometry:
                type: LineString
                coordinates:
                  - [18.0686, 59.3293]
                  - [18.0649, 59.3326]
        message: Route fetched successfully
...
    AdminUserCreateRequest:
      title: AdminUserCreateRequest
      required:
        - email
        - password
        - role
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - ATTENDANT
            - CUSTOMER
        attendantGroup:
          $ref: '#/components/schemas/AttendantGroupDto'
      example:
        email: vakt@example.com
        password: secret123
        role: ATTENDANT
        attendantGroup:
          id: 1
          name: Stockholm Södra
    AdminUserCreatedResponse:
      title: AdminUserCreatedResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserAdminDetailDto'
        message:
          type: string
      example:
        data:
          id: 7
          role: ATTENDANT
        message: User was created successfully
    UserAdminDetailDto:
      title: UserAdminDetailDto
      required:
        - id
        - role
      type: object
      properties:
        id:
          type: integer
          format: int64
        role:
          type: string
          enum:
            - ADMIN
            - ATTENDANT
            - CUSTOMER
      example:
        id: 7
        role: ATTENDANT
    AdminUsersResponse:
      title: AdminUsersResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserAdminDetailDto'
        message:
          type: string
      example:
        data:
          - id: 1
            role: CUSTOMER
          - id: 7
            role: ATTENDANT
        message: Users fetched successfully
    AttendantGroupDetailDto:
      title: AttendantGroupDetailDto
      required:
        - name
        - attendants
      type: object
      properties:
        name:
          type: string
        attendants:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/UserAdminDetailDto'
      example:
        name: Stockholm Södra
        attendants:
          - id: 7
            role: ATTENDANT
    AdminAttendantGroupsResponse:
      title: AdminAttendantGroupsResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AttendantGroupDetailDto'
        message:
          type: string
      example:
        data:
          - name: Stockholm Södra
            attendants:
              - id: 7
                role: ATTENDANT
        message: Attendant groups fetched successfully
    DeletedUserDto:
      title: DeletedUserDto
      required:
        - id
      type: object
      properties:
        id:
          type: integer
          format: int64
      example:
        id: 7
    DeletedUserResponse:
      title: DeletedUserResponse
      required:
        - data
        - message
      type: object
      properties:
        data:
          $ref: '#/components/schemas/DeletedUserDto'
        message:
          type: string
      example:
        data:
          id: 7
        message: User deleted successfully
    UserDetailDto:
      title: UserDetailDto
      required:
        - token
      type: object
      properties:
        token:
          type: string
          description: JWT token där `sub` claim innehåller användarens id.
      example:
          token: eyJhbGciOiJIUzI1NiIsInR5cCI6...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...